feat(cli): cli reads txt files given by -u flag and stores in datastore
Description: if the -u flag is used and a path argument is provided, the cli
will now attempt to read the comments of the file at that path, throwing
an appropriate exception if the file is not found or not in the correct format.
If the file is found and in the correct format, the datastore in RecordSet.txt
will be updated. This operation is idempotent, with unique records being
identified by the combination of their STB, TITLE, and DATE, if a record uniquely
identified thusly is already present and a new record is uploaded with the same
unique identifier, this record will be overwritten by the new one, whether or not
the other fields hold the same values.

Todo: This code is complex and verbose enough that it should probably be extracted
into its own module before we add any additional functionality. Also, there should
be some more error checking around reading from and writing to RecordStore.txt, and
we could probably refactor our current Try/Except blocks to include them, so that we
don't have to open RecordStore twice.